backend/requirements.txt
fastapi
uvicorn
yfinance
pandas
scikit-learn
numpy
httpx
websockets
xgboost
# PostgreSQL
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
# Auth
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart